requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-3-jet==1.0.7
django-admin-honeypot==1.1.0
django-braces==1.15.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-jazzmin==2.4.8
django-log-viewer==1.1.4
django-oauth-toolkit==1.5.0
django-oauth2-provider==0.2.6.1
django-oidc-provider==0.7.0
django-rest-framework-social-oauth2==1.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-oauth==1.1.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
kombu==5.2.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.2
Pillow==7.0.0
prompt-toolkit==3.0.22
psutil==5.8.0
psycopg2-binary==2.9.1
pycparser==2.21
pycryptodomex==3.11.0
pyjwkest==1.4.2
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
